Create README.md
This is one of my projects from the final assessment for a subject in the faculty.

This system will help to reduce the time consumed for a staff to key in and to look  into the student and parcel information. Additionally, this system serves as a platform for notifying students if their parcel has arrived at the parcel centre and will give them platform to check their parcel status. This feature will make it easier for individuals to keep informed about the arrival information of their parcel. Therefore, this proposed system is to introduce development about parcel centre, which will assist student and staff in many ways.